+++ src/Ydb.Sdk/src/Ado/YdbAdoDefaultSettings.cs
@@ -1,0 +1,12 @@
+namespace Ydb.Sdk.Ado;
+
+internal static class YdbAdoDefaultSettings
+{
+    internal const string Host = "localhost";
+
+    internal const int Port = 2136;
+
+    internal const string Database = "/local";
+
+    internal const bool UseTls = false;
+}